Organized code according to MVC architechture
